+++ cpp/include/multigrid/detail/transfers.hpp
@@ -6,10 +6,10 @@
 namespace top3d { namespace mg {
 
 void MG_Prolongate_nodes(const MGLevel& Lc, const MGLevel& Lf,
-                         const std::vector<float>& xc, std::vector<float>& xf);
+                         const std::vector<double>& xc, std::vector<double>& xf);
 
 void MG_Restrict_nodes(const MGLevel& Lc, const MGLevel& Lf,
-                       const std::vector<float>& rf, std::vector<float>& rc);
+                       const std::vector<double>& rf, std::vector<double>& rc);
 
 } } // namespace top3d::mg
